tests for products
